changed id type to Long

+++ MAS_koncowy/MAS-KONCOWY/src/main/java/edupjamas/s30338/entity/kompozycja/NormalOffice.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class NormalOffice {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int normalOfficeId;
+    private Long normalOfficeId;
 
     @NotBlank
     private String field;
