+++ MAS_koncowy/MAS-KONCOWY/src/main/java/edupjamas/s30338/entity/kompozycja/StateOffice.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class StateOffice {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int stateOfficeId;
+    private Long stateOfficeId;
 
     @NotBlank
     private String state;
