+++ MAS_koncowy/MAS-KONCOWY/src/main/java/edupjamas/s30338/entity/kwalifikowana/Application.java
@@ -17,7 +17,7 @@ import java.util.Map;
 @ToString
 public class Application {
     @Id
-    private int applicationId;
+    private Long applicationId;
 
     @NotNull
     private LocalDate data;
